+++ modules/src/ics04_channel/msgs/recv_packet.rs
@@ -88,7 +88,7 @@ impl From<MsgRecvPacket> for RawMsgRecvPacket {
 }
 
 #[cfg(test)]
-mod test_util {
+pub mod test_util {
     use ibc_proto::ibc::core::channel::v1::MsgRecvPacket as RawMsgRecvPacket;
     use ibc_proto::ibc::core::client::v1::Height as RawHeight;
 
@@ -99,7 +99,7 @@ mod test_util {
     /// proof height as well as the timeout height.
     pub fn get_dummy_raw_msg_recv_packet(height: u64) -> RawMsgRecvPacket {
         RawMsgRecvPacket {
-            packet: Some(get_dummy_raw_packet(height, 1)),
+            packet: Some(get_dummy_raw_packet(height, 9)),
             proof_commitment: get_dummy_proof(),
             proof_height: Some(RawHeight {
                 revision_number: 0,