+++ proto/src/prost/cosmos.base.v1beta1.rs
@@ -5,9 +5,9 @@
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Coin {
     #[prost(string, tag="1")]
-    pub denom: std::string::String,
+    pub denom: ::prost::alloc::string::String,
     #[prost(string, tag="2")]
-    pub amount: std::string::String,
+    pub amount: ::prost::alloc::string::String,
 }
 /// DecCoin defines a token with a denomination and a decimal amount.
 ///
@@ -16,19 +16,19 @@ pub struct Coin {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct DecCoin {
     #[prost(string, tag="1")]
-    pub denom: std::string::String,
+    pub denom: ::prost::alloc::string::String,
     #[prost(string, tag="2")]
-    pub amount: std::string::String,
+    pub amount: ::prost::alloc::string::String,
 }
 /// IntProto defines a Protobuf wrapper around an Int object.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct IntProto {
     #[prost(string, tag="1")]
-    pub int: std::string::String,
+    pub int: ::prost::alloc::string::String,
 }
 /// DecProto defines a Protobuf wrapper around a Dec object.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct DecProto {
     #[prost(string, tag="1")]
-    pub dec: std::string::String,
+    pub dec: ::prost::alloc::string::String,
 }